(migrate) ts to js
migrate typescript to javascript
