--- src/lib/types/Post.ts
@@ -1,7 +1,0 @@
-export type Post = {
-	title: string;
-	slug: string;
-	description: string;
-	date: string;
-	categories: Array<string>;
-};
