+++ src/routes/api/posts/+server.js
@@ -1,8 +1,7 @@
 import { json } from '@sveltejs/kit';
-import type { Post } from '$lib/types/Post';
 
 async function getPosts() {
-	let posts: Post[] = [];
+	let posts = [];
 
 	const paths = import.meta.glob('/src/posts/*.md', { eager: true });
 
@@ -11,8 +10,8 @@ async function getPosts() {
 		const slug = path.split('/').at(-1)?.replace('.md', '');
 
 		if (file && typeof file === 'object' && 'metadata' in file && slug) {
-			const metadata = file.metadata as Omit<Post, 'slug'>;
-			const post = { ...metadata, slug } satisfies Post;
+			const metadata = file.metadata || {};
+			const post = { ...(typeof metadata === 'object' ? metadata : {}), slug };
 			posts.push(post);
 		}
 	}
