+++ src/routes/api/search-index/+server.js
