--- src/routes/contact/+page.svelte
@@ -1,1 +1,0 @@
-<h1 class="text-2xl">Kontakt</h1>
