+++ src/routes/content/+page.server.js
@@ -1,8 +1,6 @@
-import type { Post } from '$lib/types/Post';
-
 export async function load({ fetch }) {
 	const response = await fetch('api/posts');
-	const posts: Post[] = await response.json();
+	const posts = await response.json();
 	return {
 		items: posts
 	};
