+++ src/routes/content/[slug]/+page.svelte
@@ -1,4 +1,4 @@
-<script lang="ts">
+<script>
 	import ContentItem from '$lib/components/ContentItem.svelte';
 	let { data } = $props();
 </script>
